DSR-294: move Mutation Observer into shared component
We want the absolute timestamps on all Snaps, regardless
of where they appear or what format: PDF/PNG on all SitReps,
Cards, Flash Updates.
