+++ pages/_lang/country/_slug/card/_sysid.vue
@@ -25,6 +25,7 @@
 <script>
   // Mixins
   import Global from '~/components/_Global';
+  import Page from '~/components/_Page';
 
   // Components
   import AppBar from '~/components/AppBar';
@@ -42,7 +43,10 @@
   const client = createClient();
 
   export default {
-    mixins: [Global],
+    mixins: [
+      Global,
+      Page
+    ],
 
     components: {
       AppBar,
