+++ pages/_lang/country/_slug/flash-update/_sysid.vue
@@ -27,6 +27,7 @@
 <script>
   // Mixins
   import Global from '~/components/_Global';
+  import Page from '~/components/_Page';
 
   // Components
   import AppBar from '~/components/AppBar';
@@ -39,7 +40,10 @@
   const client = createClient();
 
   export default {
-    mixins: [Global],
+    mixins: [
+      Global,
+      Page
+    ],
 
     components: {
       FlashUpdate,
